Rename 'initialiseLightDBEnclaveClient' to 'initialiseLightDBEnclave'
The method 'initialiseLightDBEnclaveClient' has been renamed to 'initialiseLightDBEnclave' across all instances within the opendsu-sdk modules. This simplifies the method naming, aligns it more closely with its actual function, and removes confusion over the role of the method. All the relevant references to this method in different modules were simultaneously updated to reflect this change.