Revise middleware behavior for apiKey and clientCredentialsOauth
The middleware handling behavior for apiKey and clientCredentialsOauth has been updated. In the apiKey middleware, the req.skipClientCredentialsOauth line has been removed. In clientCredentialsOauth middleware, the rejection process has shifted from ending the response to moving on to the next middleware. In addition, a new property, req.skipSSO, has been declared to help in skipping SSO authentication if required.