Refactor OAuthMiddleware
The diff consists of changes to OAuthMiddleware.js in the opendsu-sdk. The code was refactored for better readability and adjustments were made to how cookies are configured, including removal of the 'Max-age' property in some cookies. Additionally, a change was made to store 'logoutUrl' and 'postLogoutRedirectUrl' in local storage instead of parsing from the cookies, resulting in a simpler and cleaner approach. The change also involved the deletion of a function that parsed cookies as it was no longer necessary.