Correct typos and update record logic in fixedUrls middleware
Typos in the code comments in the fixedUrls middleware have been fixed for better clarity. Notably, 'concurency' was changed to 'concurrency'. In addition, the function for updating a record has been improved. Previously, it tried to update the record using the old 'record' object. Now, it correctly uses the 'newRecord' object for updating.